Prevent array index underflow; add tests
Cast the array index to signed before comparing to zero in DBDLLCore.cpp to avoid unsigned underflow and fix a small comment. Add comprehensive unit tests (tests/test_array_index_bounds.cpp) that validate Next/Previous/Top/Bottom/count behavior and stability, and register the new test in tests/CMakeLists.txt. Files changed: Dark Basic Public Shared/Dark Basic Pro SDK/Shared/Core/DBDLLCore.cpp, tests/CMakeLists.txt, tests/test_array_index_bounds.cpp.

+++ Dark Basic Public Shared/Dark Basic Pro SDK/Shared/Core/DBDLLCore.cpp
@@ -2991,13 +2991,13 @@ DARKSDK void PreviousArrayIndex(DWORD dwArrayPtr)
 	// dec array index
 	if(dwArrayPtr)
 	{
-		if(*((DWORD*)dwArrayPtr-1)>0)
+		if((int)*((DWORD*)dwArrayPtr-1)>0)
 		{
 			*((DWORD*)dwArrayPtr-1)=(*((DWORD*)dwArrayPtr-1))-1;
 		}
 		else
 		{
-			// First index reachable just just outside range <0
+			// First index reachable just outside range <0
 			*((DWORD*)dwArrayPtr-1)=(DWORD)-1;
 		}
 	}
